70. freelancer-rates

+++ exercism-cpp/reverse-string/reverse_string.cpp
@@ -4,7 +4,7 @@ namespace reverse_string {
 
     std::string reverse_string(std::string str) {
         auto [i, j] =
-            std::make_pair(static_cast<std::size_t>(0), str.size() - 1);
+            std::make_pair(0, static_cast<int>(str.size()) - 1);
 
         while (i < j) {
             std::swap(str[i], str[j]);
